Use bool instead of exceptions, and register singleton for repository

+++ src/Repositories/RefreshTokenRepository.php
@@ -2,7 +2,6 @@
 
 namespace Albet\SanctumRefresh\Repositories;
 
-use Albet\SanctumRefresh\Exceptions\InvalidTokenException;
 use Albet\SanctumRefresh\Helpers\CheckForRefreshToken;
 use Albet\SanctumRefresh\Models\RefreshToken;
 
@@ -20,16 +19,15 @@ class RefreshTokenRepository
         return false;
     }
 
-    /**
-     * @throws InvalidTokenException
-     */
-    public function revokeRefreshTokenFromToken(string $plainRefreshToken): void
+    public function revokeRefreshTokenFromToken(string $plainRefreshToken): bool
     {
-        CheckForRefreshToken::check($plainRefreshToken);
+        if (!CheckForRefreshToken::check($plainRefreshToken)) {
+            return false;
+        }
 
         $findTokenId = explode('|', $plainRefreshToken);
         $findTokenId = $findTokenId[array_key_first($findTokenId)];
 
-        RefreshToken::find($findTokenId)->delete();
+        return RefreshToken::find($findTokenId)->delete();
     }
 }
